refactor #27 new file for DayNightMode
